More ending params, reworked castle level end cutscenes/character json loading
Important to note that the PoseDoor behavior for castles was moved to LevelExit.gd. Aside from that, a bunch of new parameters for modifying speed during different cutscenes, a more distinct variable for determining if a character should get forced out of blocks rather than forcing the player's hitbox state, and a way to change the X offset of where pose animations will occur.