Relatively minor sound tweaks + print cleanup
The player's death sound is now split up into 2 sounds, the sting which plays on the sound track, and the usual jingle, which plays on the music track. The jingle was also upgraded to a wav for that sweet, juicy lack of delay on start.

Thwomps falling, Bob-ombs exploding and P-Switches getting pressed are all classified as their own sounds now as well, separate from "cannon".