Fixed marathon best saving logic + marathon results UI (#768)
Getting a new best in marathon mode that doesn't shave off a second of time (i.e. if you went from a time of 17.54 to 17.10) will no longer be internally ignored and now gets properly saved. Along with that, the world select menu now uses the proper marathon tracker icons.