Minor projectile optimization
Optimizes projectiles by simply performing a single if check for if the projectile has collision or not to run the surface bouncing code, which is now contained in its own area. If it does, then it will run a preliminary check for if its colliding at all before doing any of the projectile checks. Also moved bounce count to be a single time thing rather than being handled seperately by every different collision type.